pyproject.toml
[tool.poetry]
name = "compas-pro-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
psycopg = {extras = ["binary"], version = "^3.1.17"}
pillow = "^10.2.0"
django-extensions = "^3.2.3"
ipython = "^8.20.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.0"}
django-redis = "^5.4.0"


[tool.poetry.group.linters.dependencies]
pre-commit = "^3.6.0"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
